move getPost logic into Posts component
